Opening Push
Started working with a proper local and remote branch system instead of manually uploading it to GitHub. Much of the code is still a work in progress, but hopefully it'll become easier to update the project and complete tasks.
The .gitignore will need to be updated in order to ignore all the unnecessary files. The library class has not been updated at this time, and will be updated less frequently.